add alpine-3-salt-x86_64.json

+++ scripts/common/minimize.sh
@@ -14,6 +14,9 @@ if [ -n "$swap_device_uuid" ]; then
 elif [ -n "$swap_device_label" ]; then
   swap_device=`readlink -f /dev/disk/by-label/"$swap_device_label"`
 fi
-sudo /sbin/swapoff "$swap_device"
-sudo dd if=/dev/zero of="$swap_device" bs=1M || :
-sudo /sbin/mkswap ${swap_device_label:+-L "$swap_device_label"} ${swap_device_uuid:+-U "$swap_device_uuid"} "$swap_device"
+
+if [ -n "$swap_device" ]; then
+    sudo /sbin/swapoff "$swap_device"
+    sudo dd if=/dev/zero of="$swap_device" bs=1M || :
+    sudo /sbin/mkswap ${swap_device_label:+-L "$swap_device_label"} ${swap_device_uuid:+-U "$swap_device_uuid"} "$swap_device"
+fi
